Scummer is running on eu.hardfought.org
Currently scummer is set to look for Int >= 17, one poly item, and a
marker or teleport control. Scummer also quits if a wand is in the first
room (check for WoW manually).

Currently it is designed to work on hardfought with windowtype:curses
and perm_invent set, this minimises the amount of work the automated
scummer has to do.

Needs a bit of tweaking for server latency as with hdf-us the ping is >
the 40ms usleep, and commands are sent double.
There is also a hardfought/dgamelaunch issue due to overwriting ttyrecs
- disconnection occurs with the message do you want to overwrite the
  file y/n. The file names are designated by time down to the second,
scumming too fast sets off this bug. Add a one second sleep as a
workaround.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>