Add detailed usage text and cli flags for tuning
Change a lot of things quite substantially.
Add a usage text detailing required setup with screen & options
that must be set in your player rcfile, etc., things the script
assumes for simplicity and it won't work properly without.

Add a mechanism for tuning scumming behaviour using CLI flags.
Random wand in first room trumps all the other targets that can be set.

The processing of curses screendumps is still a bit crude, eventually
I'd like to have the script use the boxes to separate out the top left
messages area, the map, the status lines, and the perm_invent box over
to the right.

It works pretty well atm tho.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>